gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.1
python-dotenv==1.0.0
pydantic-settings==2.0.3
pydantic==2.5.0
